Pylance (#12)
* Type Checking Mode: basic

* Type Checking Mode: standard

* Delete: query_count

* Delete: unused

* Update: use Path

* Update: use Path

* Update: no print exp

* Update: runner

* Update: log_parser

* Update: strict partial

* Update: use super

* Update: Optional[str]

* dataclasses.asdict(self)

* Update: ABC

+++ benchmark/service/task_executor/task_execute_result.py
@@ -1,7 +1,6 @@
 import dataclasses
 
 
-
 @dataclasses.dataclass
 class StatSummary:
     """Statistical summary of a list of numeric values"""
@@ -66,7 +65,7 @@ class SingleTaskExecuteResult:
     cpu_avg_percent: float
     cpu_samples_count: int
     cpu_sampling_interval: float
-    peak_memory_bytes: float
+    peak_memory_bytes: int
     execution_time: float
     monitor_record_execution_time: float
     output_rows: int
